Local test in CI - keep all the logging
The logs are already retained as a downloadable artifact for runs
within CI, and are the only way to see any more detail about a failing
test etc - but unforunately currently only contain the final part of
the logging.

Adds --tail="all" to docker-composes log for local tests though it's
unclear from the documentation if this is the default or not.

Related docs at:

https://docs.docker.com/config/containers/logging/json-file/#options