Capture mongodb dump on completion of local tests in CI
As the local tests run against a locally deployed version of the
suite, there's no wait to view the normal test results page during or
after the test.

As a poor second best, archive off the database so it can always be
inserted into a local instance to view it.

The database seems to be about 8M in size, which is smaller than the
server.log we already capture. gitlab apparently don't seem too
concerned about the size of artifacts:

https://docs.gitlab.com/ee/user/gitlab_com/index.html#gitlab-cicd