BrowserControl: Clear text field before entering value
The login_hint test result in Authlete prefilling the user field with
'john'. The automation then enters 'john' changing it to 'johnjohn' and
hence causing authorization to fail.

I'm not sure it's currently documented that text is added rather than
replaced, so just change to a replace (which is actually what I believed
it currently did until I ran into this problem.)