Using setStatus() error if it's not making a change
This was something I noted in 800d6ceff616cb94ad792e1db6fca67f6efb53e6
as a 'problem for another day'. I vaguely knew that the RUNNING->RUNNING
could happen resulting, but that case is now handled properly so it seems
sensible to raise an exception if there are other non-change transitions
trying to happen.

I'm slightly concerned there might be mostly harmless cases where we
(say) set INTERRUPTED or FINISHED twice, which are probably harmless.
But probably best to start strict and relax the test only if necessary.