Prevent 'fireTestFinished()' running on two threads at same time
We use an atomic boolean to make sure only one finalisation block
runs, which avoids a race condition where the placeholder watcher and
the 'normal' test flow both end up running fireTestFinished() on
different threads at the same time. (Only possible when the test is
one that has a placeholder, but can also optionally complete without
the user filling the placeholder - e.g. if the OP redirects back with
an error.)

There were various choices here and this may or may not be the best
way; other approaches might have included keeping the lock whilst
fireTestFinished() ran or finding a way to kill the placeholder
watcher before fireTestFinished() marks the placeholder as satisfied.
(The fact that I can imagine  killing /other/ threads close to
the start of fireTestFinished() may well make sense in the future is
the main reason I put the changes in TestExecutionManager.)

(I saw this happening occasionally after running the previous fixes a
few times, and it pretty confusing to the end user and made the test
fail.)