run-test-plan.py: Always print all failures
In some cases run-test-plan.py would call sys.exit(1) before the end,
so you'd only see some of the problems found. Now it always runs to
the end. We also make it treat expected warnings/skips that didn't
happen as a failure, so we don't accidentally accumulate "expected"
entries that actually are no longer relevant.