Remove duplicate checks of c_hash
This log:

https://staging.certification.openid.net/log-detail.html?log=XhNQiy99c8&public=true

shows that oidcc-server is running the c_hash checks twice - once in
AbstractOIDCCServerTest when checking the authorization code, and once in
OIDCCServerTest when checking the id_token.

That's really unnecessary, so remove one set - in this case I removed
the set in the superclass on the grounds it's probably not necessary
to check c_hash in every single test module and there's a small preference
towards test modules testing specific targetted things.

spotted whilst looking at #731