Fix test failure when expires_in isn't returned
RFC6749 says expires_in is recommended, so tests should expect expires_in
to sometimes by missing, but should issue a warning if it's missing.

The @PostCondition in ExtractExpiresInFromTokenEndpointResponse was
stopping tests running, as ExtractExpiresInFromTokenEndpointResponse was
returning successfully. We just change it to returning an error, which

a) means the post condition no longer applies

b) means an warning is generated when the access token is missing
(as the calls are all "continueOnFailure" which is implicitly a warning)

The unit tests missed this as they don't differentiate from a condition
failure due to a post condition and a normal condition failure
(both are ConditionError exceptions, just with a flag).

Also fill out / correct some requirements clauses in this area.

closes #731