Raise warning if name claim returned in id_token when not requested
The condition is basically saying "I only asked for name in the
userinfo, so if the server returned it in the id_token I have no idea
why".

To explain the logic:

Returning a claim in the id_token because it was requested to be
returned from userinfo is pretty wrong, and very wrong if the id_token
was returned in the front channel (as this test isn't using encrypted
id_tokens).

Returning the name claim in the id_token by default regardless of
whether it was asked for or not (which is the other reason it could be
present in the id_token in this test) seems just as wrong - the AS
should only return information the RP requested. (Due to the general
data minimisation principle, but returning unrequested info would also
be a violation of laws like GDPR.)

(The spec does mention this, e.g. 5.1:

> This specification defines a set of standard Claims. They can be
> requested to be returned either in the UserInfo Response, per Section
> 5.3.2, or in the ID Token, per Section 2.

but I believe doesn't go as far as saying unrequested claims must not
be returned when not requested.
)