oidcc-unsigned-request-object: Handle not supported error
This test is meant to not-fail if the OP returns a
request_not_supported error, and has code to do that - but the code
is never called. Rework so that the code is actually called...

Also warn if the discovery endpoint does not indicate the request
parameter is supported, as the python test did this.