Change oidcc-client-test-refresh-token-invalid-azp to use a valid azp
It was returning an invalid azp before but what we wanted to test is
  "if no azp Claim was present in the original ID Token, one
    MUST NOT be present in the new ID Token"

So I changed it to add a valid azp value and expect clients to detect
that the id_token for the refresh request contains an azp but the original
does not.

I think most clients will fail this test and this should be optional at best.