Remove 'View running tests' link from index page
I'm not sure this serves any real use, other than perhaps for admins.
Most of what is shown overlaps heavily with the 'view all available logs'
view. I've personally not visited 'view running tests' once in 6+ months.

The actual page is still there and can be accessed if you know the url.

The rational for removing is simply that it seems better to only expose
useful things; the more things we expose the more the user has to think.