Implement OP-Rotation-OP-Sig
The python test appears to insist the JWK have "use": "sig", which is
weird as it does not insist on that anywhere else. Nor does it check
that the kid in the new key is different to the old key, or that the
old key is still present in the JWK, or that id_tokens are signed by
the new key after rotation.

Compared to python this test:

- works even if 'use':'sig' is not in the JWKS

- warns if the old key is not still present (a 'should' in the spec)

- checks that the new key has a different kid and a different public
exponent (the former because the spec says so, and the latter because
if you're still using the same private key you've not actually done
key rotation)

To implement this we start using the 'CONFIGURED' state again, and for
this one test prevent auto-start of the test, as that was the simplest
way to let the user to do something before the test continued.

Neither of the OPs used in the CI (Authlete, node-oidc-provider) have
a way for the CI to trigger key rotation, so this test fails for both
of them.

closes #474