Prevent java.util.ConcurrentModificationException during cleanup
The error shown in #738 can only happen if a new task is being added
(as that is the only place we actually add/remove from the Futures array).

I'm not 100% sure when this case might arise so I've made it throw an
exception; once we have an example of where it happens we can figure out
what the correct handling is.

closes #738