Fix race that causes incorrect lock handling during cleanup
The placeholder thread could potentially called setStatus(RUNNING)
then get killed before the setStatus(WAITING) call in fireTestFinished.

This resulted in the test incorrectly being in RUNNING and thinking it
had the lock, resulting in the cleanup tasks being run without the lock.

Refactor to avoid that errant 'setStatus(RUNNING)' being required - which
it only was because of the 'setStatus(WAITING)' at the start of
fireTestFinished().

closes #741