Improve locking around 'AbstractTestModule.status'
There was no locking around the important reads of the status in
'setStatus', which seems to be giving rise to bugs like #741.

We change the existing lock to be taken through the whole of
'setStatus', preventing any potential for race conditions.

(The obvious solution, of adding a new lock held just whilst
checking/updating status had the problem of potentially deadlocking if
that lock and the environment locks were taken in a different order.)

Note that 'getStatus()' does not take the lock (as is the case before
this change), to avoid the UI pausing waiting for the lock anytime the
test is in RUNNING status. It might be possible to add in a lock
specifically for protecting status that would only be taken during
setStatus/getStatus (to avoid deadlock you'd need to make sure to only
take the status lock *after* taking the environment lock, meaning you
couldn't use 'syncronised' on the methods). The UI would potentially
still not update during test cleanup (which is done during
setStatus(FINISHED), but that would hopefully be a smaller issue.

closes #741