Prevent test modules very occasionally being started twice
The code added in !886 could trigger for other test modules (as they
all pass through the CONFIGURED state; the test modules other than the
OP rotation one just carry on automatically) - resulting in them being
started twice.

We just make the code added to run-test-plan.py only trigger for the
test module that needs it, avoiding the problem.

Longer term I guess we may want to make it impossible to start a test
module twice, but I can't currently see a simple way to achieve that
without making changes to at least one of the configure() or start()
method that's implemented in every single test module.

closes #743