Fix ConcurrentModificationException in cleanup task
There appears to be a race condition in TestExecutionManager where it
adds to the Futures array - it does this after the relevant task has
been submitted, meaning it may have started running, which is
particularly bad in the case of cleanup where one of the first things
it does is start iterating the Futures array to cancel all other tasks.

Avoid all the race conditions entirely by making the methods
syncronized. All the access to the futures array is from these methods.

Since all the methods are now syncronized, we can simplify the code
slightly by replacing the AtomicBoolean with a simple boolean.

closes #738