Create test plan for hybrid "certification profile"
To mirror what the python does, we need to be able to run tests with
different response types and client authentication types - we go on
step further than the python and put it all in the same test plan,
meaning the user can download a single zip file that matches what they
need to submit and that an easy overview of all the tests is viewable.

The actual definition of the hybrid plan could be simplified a little,
it's been deliberately done as it has to be consistent with the order
of the tests in the certification profile pdf document. (I suspect in
the medium term we should try and generate the table in the
certification doc from the test suite directly, as that would make
what was being tested both clearer and more accurate.)

For implementation we add a new method that test plans can implement,
testModulesWithVariants(). (The differs from the current way of
defining test plans, which uses annotations and for the same result
would have been considerably more verbose / less flexible.)

As well as returning a list of test modules, this returns some presets
for variants - the only choice the user is left with is whether to use
static or dynamic client registration.

To implement this mostly means wrapping changing the list of test
modules in a plan (at the API level) from a list of strings to a list
of objects that have the old string plus a structure recording any
variants defined in the plan. When creating a test, the caller needs to
also specify which variant of the test to create if the plan uses this
new feature. (In the UI, the user just sees a big list of tests, some
of which have different variants.)

Note that this test plan currently can't be run in the CI as the
python doesn't support the new format; this doesn't significantly
lower code coverage as we still have the plain oidcc test plan that
tests all the test modules in various configurations. It would be good
to add support to the python in due course though.

closes #678