Prevent testplan creation if a required variant is missing
Currently if the user tries to create a testplan using the API but fails
to specify all the variants, the testplan may be successfully created
but only contain one of two test modules. This has (a few months ago)
been seen happening in the CI when new variants are added but we forget
to update scripts for running the test to specify the new variants.

Instead we fault if a test module finds one of it's variants has not been
specified.

(The UI already prevents this happening as it forces selections for all
variants before allowing the user to create a test plan.)