Add config certification profile testplan
A slightly odd plan as it only contains a single test module, but that's
all that is necessary.

On review of the PDF a few tests were missing from the discovery test
module (some because they were tested in other modules/profiles, some
because the python test was, on review, more comprehensive than the
existing java one); these have been added/tweaked as necessary.

One test mentioned in the certification pdf is irrelevant and also not
included in the python.