Add form post certification profiles
As per:

https://openid.net/wordpress-content/uploads/2018/06/OpenID-Connect-Conformance-Profiles.pdf

The form post tests must be run for each of basic/implicit/hybrid that
are supported.

There's a few different ways this could have been done, as per
discussion at:

https://gitlab.com/openid/conformance-suite/-/merge_requests/893#note_336533544

We've gone for creating 3 separate test plans, one for each of the
underlying profiles.

(Trying to give the user a way to create one plan which optionally had
each of the 3 underlying profiles would possibly be slightly nicer but
would also require more changes to the core of the variant system.)