Add support for static server configuration
This allows servers that don't have a discovery endpoint to run and
pass the basic/implicit/hybrid certification test plans, including the
form post versions.

We deliberately don't offer support for non-discovery-supporting
servers in other cases as the other certification profiles (config,
dynamic) require the server to support discovery as per:

https://openid.net/wordpress-content/uploads/2018/06/OpenID-Connect-Conformance-Profiles.pdf

The majority of the changes are to add configuration fields in the
front end for the necessary values (issuer, jwks_uri,
authorization/token/userinfo endpoints) - with further changes to make
sure we only show these when the user selects static configuration
(and to hide the discovery url in that case).

Overall there is a focus on keeping the configuration forms as simple
as possible - avoiding presenting values that the user does not need
to consider for the tests they will run. (The implicit profile
configuration forms, for example, contain 11 entries in the java
version, compared to over 5 screenfuls in python.)

Small changes were necessary to GetDynamicServerConfiguration /
GetStaticServerConfiguration as although in theory both worked, in
practice they threw errors.

Various places that check discovery information is correct for the
particular thing being tested are guarded to only run when the server
supports discovery.

closes #726