Fix 'Repeat Test' button on log-detail not working in some cases
The problem was that we weren't finding the test module to update,
because we were looking for an entry with more variants fixed by the
plan than was possibly true.

The API had been assuming that all variants passed to the API are
ones that are fixed in the plan definition, which wasn't the case here.

Now we explicitly remove the variants that were user selected, leaving
only the variants that are from the plan definition. (Or, if someone
has called the API incorrectly there may be other variants passed and
the updateTestPlanWithModule will fail, correctly/expectedly.)

closes #750