Fix false failure in fapi-ciba-id1-ensure-mtls-holder-of-key-required
If the AS dropped the connection due to the missing TLS client certificate
(which is one of the allowed outcomes as many TLS terminators don't
allow more graceful errors), then we still check the content-type, which
of course won't be set.

The content type is already checked in (the rather misnamed)
validateAuthorizationEndpointResponseSteps call a few lines below, so
we can just remove this one.

closes #765