Lock local version of mongodb to 4.2
As per #754 we're currently using differing versions of mongo in the
cloud vs when developing locally.

Developers (and anyone deploying using docker-compose) are currently
running 4.2 locally, as that's the latest. Lock us to that version so
the problem doesn't get worse. (I don't think it's wise to force a
downgrade to 3.8 as mongo doesn't guarantee that downgrade works.
Equally we can't jump the cloud upto 4.2 in a single step as mongo
doesn't guarantee a jump that big will work.)