FAPI-RW: Add support for encrypted id_tokens
Encrypted id_tokens are required by Australian Consumer Data Rights
ecosystem.

This is basic; it allows the test to run if the OP returns encrypted
id_tokens, but doesn't do any checks on the id_tokens - for example it
won't flag if the kid is wrong in the received id_token or if the
id_token is encrypted with RSA1_5 (not permitted under FAPI).

As it's all quite generic it might mean FAPI-CIBA & OIDCC tests now
supported encrypted id_tokens too, but I've not tried.

It explicitly doesn't currently support symmetric encyption, seem comments
in ExtractIdTokenFrom{Authorization,Token}Response.java