Change post/post requirements to give users better error messages
It's really not clear to users that if these checks fail it can be
a bug in the test suite. It might be something the user has done wrong,
but in many/most/all cases where the user does something wrong we'd
want to give a more explanatory error of what they might have done
wrong than a failed precondition.

Reported by Hans.