FAPI-RW-OP: Rework handling of resource endpoint headers
These were handled a bit oddly, and it wasn't really obvious from the code
what was actually currently happening.

The root cause is FAPIGenerateResourceEndpointRequestHeaders, which boths
creates a new empty set of headers, and adds the (optional)
x-fapi-auth-date header to those headers.

Various other functions would also create a new set of headers if one
didn't exist.

The code that attempted to have the second client not send an
x-fapi-auth-date header didn't call FAPIGenerateResourceEndpointRequestHeaders
and hence actually ended up reusing the headers from previous calls,
which had an x-fapi-auth-date.

The new model we split FAPIGenerateResourceEndpointRequestHeaders in two,
meaning there's an explicit function to create fresh empty headers.

We simply the functions for adding headers so they never create a new set
of headers.

We also move some lines on test initialisation that set headers that
would only be used in the ob pre-authorization intent creation to be
in the pre-authorization steps.

The reason for changing is that we need to send an extra header in the
Australian CDR case, and inspecting the code revealed this oddity
which seemed better to fix first.

This is intended to result in one actual behaviour change - the second
client now does not send x-fapi-auth-date, as was always intended.