FAPI OP: Cleanup handling of x-fapi-financial-id
This is required only for some OB UK tests, as an OB spec erroneously
includes it even though it was removed by FAPI sometime ago.

Change it to only be exposed in the configuration form for the OBUK
variant, and do not require the user to provide a value for it.