FAPI & FAPI CIBA OP tests: Test resource calls without interaction id
The original OB tests tried sending calls to the resource server both
with and without an x-fapi-interaction-id - as it's optional for the
client to send this.

However this was done in the OB specific pre-authorization steps, and
hence this test wasn't carried out in the pure-FAPI tests.

Hence change the second client to not send the header (and to expect
the AS to return an interaction id the AS creates).