FAPI-RW: Remove check for request_uri support
I think at some point FAPI had recommended servers include request uri
support, but there's nothing in the current specs that say this so
remove the clause.

The current implementations draft says:

> may support the request object endpoint as described in section 7;

and the latest in-progress draft says:

> may support the pushed authorization request endpoint as described in PAR;

Neither are recommend or should, so there's no need to make the check.