Change admin authentication to use gitlab
This allows us to manage admin access permissions in the gitlab ui,
in the same way we manage admin access to the git project etc.
It also means we can be a bit more selective than the old approach of
allowing an entire google domain access, and gitlab allows us to force users
in the group to use 2FA.

Sadly gitlab doesn't tell us which the user's access level to a group is
(see:
https://gitlab.com/gitlab-org/gitlab/-/issues/21662#note_215135363
) so we have to create a brand new group just to control admin access.

The old google way is left in place, in case people's local installs of
the suite are using that feature.