CDR: Correct acr claim
As per:

https://consumerdatastandardsaustralia.github.io/standards/#levels-of-assurance-loas

level 2 is all that's necessary for read operations, and the mandated
authentication method doesn't actually allow for LoA3 currently.

ID2 of FAPI has some text requiring clients to request acr=LoA3 with
essential=true. That text has been deleted for ID3. It’s considered
fine for clients to request a lower LoA and is at the AS’s discretion
what it level it actually applies based on the client/what the client
wants to do. (That change was made as a result of feedback in the UK,
where in some situations a lower LoA was permitted.)