Reword the various text about scope=email behaviour
Multiple users were struggling to understand this, for multiple different
reasons. Include more explicit references to the spec, the expected
behaviour, and why unexpected behaviour could have unforeseen privacy
implications.

As some people do rely on id_tokens containing email/other info (and in
some cases scope=email might be the only way to request this, if the OP
doesn't support the non-mandatory-for-most-profiles claims parameter),
make the wording slightly less certain that it's a bug.