Add CDR sharing_duration claim to request object
As reported by Andrew Latham@Ping, this claim is necessary for CDR
compliant servers to return a refresh token and hence for the refresh
token tests to actually test refresh tokens.

Note that we add it in all cases. It would probably be equally valid to
add it only for the refresh token tests.