OIDC/FAPI OP: Add test that verifies valid PKCE succeeds
FAPI-RW-ID2 does not require servers to support PKCE, but as per
https://tools.ietf.org/html/rfc6749#section-3.1 'The authorization
server MUST ignore unrecognized request parameters' - i.e. whether the
server supports PKCE or not, a valid PKCE request must succeed. The
reason for this test is that many OpenID Connect clients speculatively
use PKCE, and the OAuth2 standards require that requests from such
clients must not fail.

A similar test is added to the OIDC basic+hybrid profiles.

closes #778