Remove requirement to use dev-branch-x branchs for CI
Uses dynamic environment urls that were added to gitlab in version 12.9
(about 3 months ago):

https://gitlab.com/gitlab-org/gitlab/-/issues/17066

We need a chunk of code to deal picking which of the 9 review app
environments is used (by querying the gitlab environments API), but
otherwise it's relatively straightforward.

(The underlying reasons to use a fix set of review apps still exist -
the gitlab/google login require pre-registration of the redirect_uris,
and some of the tests against Authlete use statically configured
clients where the redirect uris have the current set of review apps
registered.)

There is a race condition here if two deployments run at once, they
might both try to claim the same url. (If that happens, deleting the
branch or stopping the environment then repushing or running a new
pipeline will fix it.)

relates to #358