FAPI-RW/CIBA: Send x-fapi-customer-ip-address
The values are the unroutable 'reserved for documentation' values used
as examples in the lastest revision of the FAPI-R spec.

There is a potential downside that banks may reject the addresses,
either because they know they're not real, or because they differ from
the one the customer actually used. Hopefully that's not the case, as
we don't currently have access to the tester's IP address, and even if
we did we need both IPv4 and IPv6 addresses.

closes #761