Update webfinger code to handle non-aliased tests
This is technically ambiguous - in the normal use of an alias, an
alias is always prefixed by 'a/', and tests without aliases have no
prefix.

The chances of a clash between an alias and a test id are hopefully
very low, so I don't imagine this is an issue.

Also,return strings with the problem when returning an error before
it manages to find the test id, and log incoming calls to the console
just in case.

closes #842