Use dotenv for conformance API token + remove environment: from test jobs
Means they're no longer treated as deployment jobs - this was
previously necessary so they could access the cluster to retrieve the
API token from mongodb, but now we use the dotenv artifact to do that.
(This will stop gitlab getting confused and thinking that deployment
failed because a test job failed.)

Also updates run-test-plan.py to cope if the url is missing the final /,
as gitlab seems to generate urls like that a lot, and it seems better to
tolerate it.

The trailing base64 padding ('=') is removed from the CI-generated API
token, as it's not allowed in dotenv, and doesn't serve a purpose anyhow.

Also removed some out of date comments.

The deployment and test jobs can now all be combined together, as
there's no longer any difference between the review apps jobs and other
environments - with the advantage that the demo/staging/production test
jobs are now split into ciba/server/client jobs (as review app tests
already were) and hence will complete faster.