FAPI-RW: Simplify PAR request object code
We were using three different methods of customising the par request
object in the test module subclasses:

1. overriding createParAuthorizationRequestObject() to add conditions
to the sequence

2. overriding createParAuthorizationRequestObject() to call conditions
before

3. overriding createParAuthorizationRequestObject() to call a different
variant of makeCreatePARAuthorizationRequestObjectSteps()

4. overriding makeCreatePARAuthorizationRequestObjectSteps() to add
to the sequence

Replace this all with method 4, meaning createParAuthorizationRequestObject
can be inlined as it's always a call to
makeCreatePARAuthorizationRequestObjectSteps() now.

Also rename AddBadRequestUriToRequestObject to
AddBadRequestUriToAuthorizationRequest, which better reflects what it
actually does.