FAPI-RW: Create a step that purely creates a request object
This is work towards #818 - by having a step that just creates the
request object (without creating the redirect too) test modules are
able to customise the request object in the same way in the PAR
and non-PAR tests.

This step hasn't done the necessary renaming, nor the commoning up of the
PAR and non-PAR code paths.