Catch Error exceptions in more places
I saw a 'ClassNotFound' error from a Selenium wait failing that only got
thrown on the console as Selenium runs on a background thread.

As we did in 4a41498840c41a9bd89bcedc35b3e7ad4ff7d8cb, catch Error as
well as Exception in TextExecutionManager so that the exception gets
recorded into the test log (and causes the test to go to interrupted
instead of sitting around in 'Waiting').

I went through and looked for other places where we catch 'Exception',
and added 'Error' to the catch as well if they looked like they fell into
the same category.