FAPI-RW: Make PAR use same request uri redirect builder as oidcc
The different behaviour of the PAR one was causing various existing
FAPI tests to not behave as were intended when running with PAR enabled.

The PAR tests relied on the behaviour of the current one, so change
that test to explicitly add the client id to the authorization request
instead.

Also makes the OIDCC tests put request_uri as a string in the
root of the environment, so that all tests have it there.