Improve error messages when problems calling out over http
Currently there's a number of circumstances (timeouts, ssl errors, etc)
where we end up presenting the tester with java stack traces, which is
not too user friendly.

Try and cover everywhere we use rest template, catch RestClientException
and use it to present a friendlier error message. (The stacktrace
will still be accessible in the 'more' section of the log entry, just
in case.)