FAPI-RW PAR: Refactor code that adds authentication a little
Remove 'BuildMTLSRequestObjectPostToPAREndpoint' as it's not really
necessary to have a MTLS specific one - make
BuildRequestObjectPostToPAREndpoin only do the basic build, and move
the code that was in it to add the client authentication assertion to
a new function.

Also move the 'build' to be called from a different place, as the
function it was in wasn't named in a way that would make you think it
would build the request.