OIDCC: Allow basic/implicit/hybrid OPs to not support request_uri
This appears to have been a mistake in the java suite; I believe the
intention in the basic/implicit/hybrid profiles is that request_uri
should either be supported, or can be rejected as not supported as
per the spec.

We had a discussion about this with the connect working group, and it
seems the request object by value tests were updated for that decision,
but not the request_uri tests.

See https://bitbucket.org/openid/connect/issues/1167/required-certification-behaviour-for

The existing request_uri test is kept (and a new one where support is
optional added), as OPs are required to support request_uri for the
dynamic profile.