FAPI-RW: Include nbf in second client request object
nbf is valid (but not required) here in FAPI-RW ID2 so we should
probably try sending it, as no server should rejects request objects
that include nbf.

nbf is mandatory here in FAPI1-Advanced, so this change helps ensure there is
a smooth migration path from FAPIRW-ID2 to FAPI1-Advanced-Final.

FAPI working group was notified of change here:

https://bitbucket.org/openid/fapi/issues/363/proposed-change-to-fapi-rw-id2

and no objections were received.