Fix for a memory accumulation
If the browser automation throws an exception, it seems to stay around
in memory either forever or for long enough that if it happens in a
few tests the JVM hits the memory limit and dies.

Remove it from the runner list in all cases where it exits, which
stops the jvm dieing.