Allow static servers to skip unsigned request object tests
This is necessary as our test server (Authlete) now aligns to jar by
default and hence the unsigned tests in the CI were failing as they
were being rejected. This allows the CI to continue testing the other
tests, and potentially allows the options for other OPs to pass
in a similar way - though I think generally we'd encourage them to
support discovery instead, Hence I've not exposed this in the
configuration form.