Workaround for occasional WebRunner timeouts seen in CI
There's a common intermittent problem with the CI where Selenium fails
to realise everything is complete. This appears to be some kind of
complex threading issue in htmlunit where the XMLHttpRequest completion
block fails to run in rare cases.

To workaround this, just assume it has completed if 5 seconds have
elapsed.

Hopefully eventually it will be fixed in htmlunit.

See https://gitlab.com/openid/conformance-suite/-/issues/766