FAPI: Remove prompt=login from 'user rejects authentication' test
prompt=login was originally added to fix #526 - and it was only added
for the pure-FAPI tests, not for the OB UK variants.

As per recent reports from Australia, it unfortunately has the side
effect of changing exactly what is being tested, introducing the extra
confusion that prompt=login says this in the OIDCC spec:

> The Authorization Server SHOULD prompt the End-User for
> reauthentication. If it cannot reauthenticate the End-User, it MUST
> return an error, typically login_required.

I'm not really sure exactly what this text means, but to avoid any
argument simply remove prompt=login making it unambiguous that
the correct response is the access_denied error.

Fix #526 in a different way by documenting that the user may need to
clear cookies.