oidcc-registration-jwks-uri should only run for private_key_jwt
This test was accidentally running for clientauth=none, which it shouldn't
as the test only tests something if private_key_jwt client authentication
is in use. Just exclude none.

This is the problem originally reported in #856 - but there are more
problems (as noted on that issue), as public clients still aren't
properly supported by the tests.