Refactor cache header verification out of BC logout condition
We have two different sets of code attempting to do this validation,
both of which suffer from bug #858 - i.e. not coping if the server
passes the cache-control header multiple times, as is allowed.

The two requirements are also slightly different, further commits
will resolve this. This commit just moves the code without changing
any behaviour.