Add error handler for failed logins
We've had a few questions about webfinger over the last year or so, and
the fact that it doesn't even show an error just adds to confusion.

Add a trivial error handler that displays a message. It still might not
be a very helpful message but it's better than nothing...

There was some existing code to show an error in login.html but I'm
not sure how that could ever be triggered.