Fix two test modules having name 'oidcc-request-uri-unsigned'
This is the reason the compare_local job in the gitlab pipeline was
sometimes showing diffs in this test - it was presumably undefined
which of the two tests with the above name would run.

The now-actually-run test is skipped as Authlete doesn't support
unsigned jwt, so add it to the expected skips list.