Add ArchUnit rule to prevent use of JSONElement getAs*
Issue #398 added a SonarQube rule that should be preventing this, but
it seems stopped working at some point.

We only use Sonarqube for this one purpose, and it's a bit of a pain
to setup / use / integrate with gitlab / add our custom rule to / etc,
so rather than attempt to fix it add ArchUnit instead which is a lot
simpler. Unfortunately I wasn't aware of ArchUnit at the time, nor
was I aware of what a pain SonarQube would turn out to be, when we
decided on SonarQube here:

https://gitlab.com/openid/conformance-suite/-/merge_requests/434#note_151671203