Remove SonarQube
It was only here to implement the check required for #398 which have
now been replaced by ArchUnit rules.

Unfortunately there's still one match when you search for SonarQube in
the source tree, which is SONARQUBE_BOT_API_KEY which is a key to
access the gitlab API which is also used by the code that finds a free
dev environment to deploy to.

(Once this is deployed we can remove sonarqube from our kubernetes cluster
and our deployment instructions.)