Remove CheckMatchingStateParameter condition
This condition duplicates CheckStateInAuthorizationResponse, except:

1) some error messages are harder to understand

2) it doesn't have a special case for invalid_request_object errors
(the special case would do no harm in the cases where
CheckMatchingStateParameter - and the special case is not the reason
the duplicate was created, it seems to have been accidentally created
during a refactor when a larger condition was split down into smaller
parts).

Hence replace calls to it with CheckStateInAuthorizationResponse, and
remove the duplicate code and corresponding unit test. (I also
checked that the code being retained passed the unit test for the code
being removed.)

This was spotted by chance whilst working on the code to check the
'iss' parameter.