Add kubernetes deprecated API mapping job
Existing deployments can't be upgraded as they were deployed using
APIs that kubernetes 1.16 (which google upgraded our cluster to) no
longer supports.

You'd think we would follow:

https://docs.gitlab.com/ee/topics/autodevops/troubleshooting.html#error-unable-to-recognize--no-matches-for-kind-deployment-in-version-extensionsv1beta1

but as we've customised our .gitlab-ci.yml quite heavily that method
won't work. Some of the customisation/reasons is listed here:
https://gitlab.com/openid/conformance-suite/-/wikis/gcp-deployment

So manually add a call to do the upgrade using a handy kubectl plugin.

This appears to be upgrading our ingress to a newer version. However
kubectl itself (and the gcp console) seem reluctant to show the newer
version, and kubectl convert fails to upgrade the ingress:

$ kubectl convert -f ./chart/templates/ingress.yaml --output-version "networking/v1beta1"
kubectl convert is DEPRECATED and will be removed in a future version.
In order to convert, kubectl apply the object to the cluster, then
kubectl get at the desired version.
error: networking.Ingress is not suitable for converting to "networking/v1beta1" in scheme "k8s.io/kubernetes/pkg/api/legacyscheme/scheme.go:30"

This seems to be some kubernetes oddity that others have also run into:

https://github.com/kubernetes/kubernetes/issues/94761

As the currently used ingress API won't be removed until kubernetes
1.22 I think we can ignore this for now.

closes #876